Paths() support filepath.ListSeparator
